requirements.txt
PySide6
./whisper
ffmpeg-python